Port core engine .c files
Layer PaperBoat port-specific changes on top of dx's improved codebase
  across 41 source files in 4 tiers:

  Tier 1 - Heavy rewrites (10 files): model.c, hud_element.c, msg.c,
  entity.c, item_entity.c, imgfx.c, animator.c, sprite.c, effects.c,
  collision.c. OTR texture/shape/asset loading replacing ROM DMA and Yay0
  decode. intptr_t widening for 64-bit. FrameInterpolation recording.

  Tier 2 - Moderate changes (9 files): main_loop.c, cam_main.c,
  draw_box.c, background.c, main.c, npc.c, 16C8E0.c, worker.c,
  101b90_len_8f0.c. nuGfxTaskStart removal, libultraship color/depth
  image setup, CALL_EVENT hooks, sprite loader integration.

  Tier 3 - State files (5 modified): state_logos.c, state_battle.c,
  state_pause.c, state_file_select.c, state_title_screen.c. OTR shape
  and texture loading, nuGfxCfb sync bypasses, DMA_COPY_SEGMENT removal.

  Tier 4 - EVT/audio/filemenu/pause (17 files): CALL_EVENT hooks for
  music/ambience/script systems, intptr_t on function pointer and message
  casts, INCLUDE_IMG replaced with OTR asset headers, OTR sprite shading.

  All dx improvements preserved: profiling, debug menu, named variables,
  surface effects architecture, nodeCache, 3-bucket render tasks.

+++ src/audio/bgm_player.c
@@ -967,7 +967,7 @@ void au_bgm_load_phrase(BGMPlayer* player, u32 cmd) {
                         track->firstVoice = linkedTrack->firstVoice;
                         track->lastVoice = linkedTrack->lastVoice;
 
-                        track->bgmReadPos = (track->bgmReadPos + (s32)player->phraseStartPos);
+                        track->bgmReadPos = (track->bgmReadPos + (intptr_t)player->phraseStartPos);
                         track->delayTime = 1;
 
                         track->linkedTrackID = linkedID;
@@ -985,7 +985,7 @@ void au_bgm_load_phrase(BGMPlayer* player, u32 cmd) {
                     curVoice += count;
                     track->lastVoice = curVoice;
 
-                    track->bgmReadPos = (track->bgmReadPos + (s32)player->phraseStartPos);
+                    track->bgmReadPos = (track->bgmReadPos + (intptr_t)player->phraseStartPos);
                     track->delayTime = 1;
                 }
             } else {
